The small changes

+++ format.h
@@ -3,5 +3,5 @@
 typedef struct
 {
  char signature[4];
- unsigned long int length:32;
+ unsigned int length:32;
 } service_information;
